# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "File Manager API",
        "description": "API Documantations Service",
        "contact": {
            "email": "varynkiv@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "API host"
        }
    ],
    "paths": {
        "/files": {
            "get": {
                "tags": [
                    "File Manager"
                ],
                "summary": "File lists",
                "operationId": "file.index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "File Manager"
                ],
                "summary": "File Store",
                "operationId": "file.store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "tags": [
                    "File Manager"
                ],
                "summary": "File Info",
                "operationId": "file.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "File ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "File not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "File Manager"
                ],
                "summary": "File Delete",
                "operationId": "file.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "File ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "File not found"
                    }
                }
            }
        },
        "/files/download/{id}": {
            "get": {
                "tags": [
                    "File Manager"
                ],
                "summary": "File Download",
                "operationId": "file.download",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "File ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "File not found"
                    }
                }
            }
        },
        "/files/total": {
            "get": {
                "tags": [
                    "File Manager"
                ],
                "summary": "File Total Used Space",
                "operationId": "file.total",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}